Revert "Fully support readOnlyRootFilesystem for each deployed container"
There are issues running with this configuration when using the
default "restricted" SecurityContextConstraint in an OpenShift
environment. Rather than applying this configuration in a patch
release, this will be addressed in a newer version.

While this patch does work in essentially every other environment,
in order to maintain maximum compatibiltiy with the default SCC,
it makes sense to revert it.

This reverts commit 000f58c1e755dfc525141afc571f6a08a4eeb062.