K8SPG-207: Add verifyTLS param to Pgtask/restore
It's also a problem in upstream. `pgbackrest/create-replica.sh` script try to detect the repo type using a file in `/pgconf` but these file only exists in case the cluster in standby mode. So it fails to detect the repo type is s3 and doesn't add the required flag to skip TLS verification.

We'll handle `backrest-s3-verify-tls` parameter in restore like Pgtask/backup to control the behaviour on restore level. If there's no parameter the operator will try to detect it itself using the pgcluster spec.