K8SPG-207: Add verifyTLS param to Pgtask/restore (#269)
* K8SPG-207: Add verifyTLS param to Pgtask/restore

It's also a problem in upstream. `pgbackrest/create-replica.sh` script try to detect the repo type using a file in `/pgconf` but these file only exists in case the cluster in standby mode. So it fails to detect the repo type is s3 and doesn't add the required flag to skip TLS verification.

We'll handle `backrest-s3-verify-tls` parameter in restore like Pgtask/backup to control the behaviour on restore level. If there's no parameter the operator will try to detect it itself using the pgcluster spec.

* K8SPG-207 Backup/restore via self-signed tls test case

* Update e2e-tests/demand-backup/run

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>

* K8SPG-207 Remove obsolete line

* check repo type before adding no-verify-tls flag

Co-authored-by: cap1984 <ivan.pylypenko@percona.com>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>