K8SPG-217: Fix missing primary deployment after restore
The problem only occurs if primary is failed over to a replica before
the restore.

The operator stores the current primary in an annotation of `pgcluster`
(i.e. `current-primary: cluster1`). After the bootstrap finishes,
operator creates the primary deployment using this value as the
deployment name. After the primary is created, it proceeds to create
replica deployments.

If a failover is performed and a replica deployment is promoted to
primary, the operator updates the annotation (i.e. `current-primary:
cluster1-repl2`).

If a restore is started at this moment in the cluster:

1. Bootstrap finishes
2. Operator creates `cluster1-repl2` as the **primary** deployment
3. Operator creates `Pgreplica` objects, `cluster1-repl1` is created but
   `cluster1-repl2` is skipped since it already exists

This commit updates the `current-primary` annotation with cluster name
just before the restore, so restore is performed on PVC `cluster1` and
replicas are created as expected.