K8SPG-233 Add fallback source of variables
In order to cut down the anount of puzzling variables
the additional defaults source has been added to ansible preflight
routine.
As a result operator.yaml contains only operator related variables.
If needed the user is able to bring some variables back, they are to
take precedence over the defaults