K8SPG-304: Don't depend on trust auth method for PMM (#398)
* K8SPG-304: Don't depend on trust auth method for PMM

We were depending on trust auth method in pg_hba to authenticate PMM
exporters to the database. It was problematic security wise and we did
it to move v2.0.0 forward. This commit fixes the issue.

We're reverting peer auth method for local connections and adding new
HBA rules for PMM monitoring user. I added a new package called `pmm` in
the internal directory to include the rules with minimal changes to
upstream code (like we did for pg_stat_monitor).

With these changes, we don't use the postgres user but create a separate
user called `monitor` to use in PMM client. That way we can ensure the
secret is created for the user, since the secret for `postgres` user is
only created if user doesn't define any users in their cr.yaml.

Also, we're now using TLS for PMM client connections.

* fix monitoring test

* fix unit tests

---------

Co-authored-by: Viacheslav Sarzhan <slava.sarzhan@percona.com>