fix `deleteAnnotation(naming.PGBackRestBackup)`
This commit changes the `deleteAnnotation` funciton. Now it will return
the `true` value only when the annotation will be deleted from the
crunchy's cluster.

After deleting the `naming.PGBackRestBackup` annotation operator waited
for 5 seconds and reconciled again, assuming that `Reconcile` was
already called for the crunchy's cluster. But on the next reconcile it's
possible that crunchy still has the annotation we want to be deleted.
This caused an additional backup per pg-backup resource.