K8SPG-518: Add latest restorable time field to backups (#759)
* K8SPG-518: Add latest restorable time field to backups

These changes add a new field called `latestRestorableTime` to
`PerconaPGBackup` objects.

Since PG Operator doesn't reconcile objects if they're not changed, we
needed introduce a new mechanism to trigger reconciliation: We create
and store a channel in both `PGClusterReconciler` and
`PGBackupReconciler` and send events to it to trigger reconciliation of
`PerconaPGBackup` objects.

When a cluster is reconciled for the first time, operator starts a
separate goroutine which continously fetches commit timestamps from
latest uploaded WAL file. If the timestamp differs from
`status.latestRestorableTime` field of the last backup, it sends an
event to the channel and since `PGBackupReconciler` watches this
channel, it reconciles the backup.

When a cluster is deleted, operator stops the goroutine by sending a
`DeleteEvent` to an additional channel called `StopExternalWatchers`.
For this we introduce a new finalizer called `percona.com/stop-watchers`.
That finalizer ensures the watcher goroutine is stopped and watcher is
removed from the watcher registry. **Users don't need to add this
finalizer to their CRs.** Operator will automatically patch the CR if
this finalizer doesn't exist in `metadata.finalizers`.

* address review comments

* fix logger panic

* address review comments

* fix

* fix pitr

* fix pitr

* remove unused code

---------

Co-authored-by: Viacheslav Sarzhan <slava.sarzhan@percona.com>